#12 refactor: struct subview to viewbuilder
plus adding UIScreen extension following Cozy's UIScreen+Extensions file

According to code review: https://github.com/DeveloperAcademy-POSTECH/MC2-Team15-MeetingNMusae/pull/19#issuecomment-1152866389
